module4/mwe/airflow-dag/requirements.txt
# Airflow DAG MWE - Minimal Dependencies
# Install with: pip install -r requirements.txt

apache-airflow==2.8.1